### 2018-09-29 #987 @whedon
[PRE REVIEW]: PyGeoPressure: Geopressure Prediction in Python
**Submitting author:** @whimian (<a href="http://orcid.org/0000-0002-4109-6358">Hao Yu</a>)
**Repository:** <a href="https://github.com/whimian/pyGeoPressure" target ="_blank">https://github.com/whimian/pyGeoPressure</a>
**Version:** v0.1.8
