### 2018-10-02 #987 @whimian
Hey @kthyng ! 💭Maybe @WalBytes from the reviewer list. It seems that he's interested in Geophysics and Petroleum Exploration.
