### 2018-10-02 #987 @WalBytes
Yes I can review this.
