### 2018-10-02 #987 @kthyng
Great, thanks all! I will add you two as reviewers, and close this issue. The official review will take place in a newly-opened issue.
