### 2018-10-05 #992 @JesperDramsch
Good to go from my side @kthyng. It's a very nice package. Works as described and is well documented. Pleasure to review.
