### 2018-10-05 #992 @WalBytes
Very useful package. The software functions as described. A minor revision in the second line of the paper; the author can re-word as follows: "The pore fluid pressure affects the physical properties of reservoir rocks, hence predicted pore pressure is a key input when building the geomechanical model of a reservoir. "
Good to go!
