### 2018-10-06 #992 @whimian
@WalBytes I've revised the paper in the latest commit.
