### 2018-10-06 #992 @WalBytes
Awesome! Good to go from my side @kthyng <https://github.com/kthyng>.
