### 2018-10-08 #992 @kthyng
Excellent @JesperDramsch and @WalBytes!

Nice work @whimian! Please create an archive (on Zenodo, figshare, or other) and post the archive DOI in this issue.
