### 2018-10-09 #992 @whimian
@kthyng  Thank you for all your assistance.

DOI for the latest release is 10.5281/zenodo.1452001
