### 2018-10-11 #992 @whimian
hey @kthyng,

There is something wrong with the author name .

My family name is "Yu", But it's "Hao" on the paper.

Could you please correct the author name?
