### 2018-10-12 #992 @whimian
@arfon I've changed the author name in `paper.md`, please recompile it.
